Refactor media-strip: core/shell split, parse-based IDs, review cleanups
Applies the findings from a full code review of packages/ui/media-strip:

- Split the package into a pure, React-free core/ (types, validation,
  collection-ops, dnd, project-validation, utils + their tests) and a
  React shell at the package root, making the architecture legible from
  the filesystem.
- Replaced all `as TimelineItemId` / `as CollectionId` casts with parse
  functions: new asTimelineItemId/asCollectionId parse-or-throw helpers
  for authoring-time-trusted IDs, parse* Result variants for untrusted
  input (DOM attributes, ToggleGroup values). Only two deliberate casts
  remain, in validation tests that exercise invalid-ID rejection.
- Curated index.ts exports; internal DOM attr constants, memo
  comparators, and dnd encoders are no longer public API.
- Deduplicated the base validation checks the video validator had
  inlined; bad collectionId now fails as "empty-collection-id" instead
  of masquerading as "empty-id".
- Perf: ancestor checks on the dnd autoscroll hot path now walk a
  memoized parent map (O(depth)) instead of scanning the item lookup.
- Keyboard adjacency uses the ordered visibleCollectionIds prop as
  canonical, with the DOM query (dead sort removed) as fallback.
- Removed dead code: unused maxScrollLeft state, SET_DRAG_WIDTH /
  SET_NEST_TARGET reducer actions, onMoveItem memo prop; the areEqual
  whitelist is now compiler-enforced via `satisfies`.
- Drag-scroll listeners now clean up via one AbortController per drag;
  frame-duration constant shared between gesture tracking and inertia.
- Split the 1,438-line types test grab-bag by module; ops/graph test
  fixtures now go through the smart constructors instead of `as any`.
- Stories: extracted pointer-simulation and fixture helpers, removed
  ~12 copy-pasted decorators that double-wrapped the canvas, and
  rebuilt ReorderDemo on the package's own applyTimelineItemCommand.

Verified: 101 unit tests + 199 Storybook browser tests pass; package
typechecks clean.
