+++ packages/ui/media-strip/core/media-strip.collection-ops.ts
