+++ packages/ui/media-strip/core/media-strip.dnd.ts
