+++ packages/ui/media-strip/core/media-strip.project-validation.ts
