--- packages/ui/media-strip/media-strip.types.test.ts
@@ -1,1438 +1,0 @@
-import { describe, expect, test } from "vitest";
-import {
-  parseTimelineItemId,
-  parseCollectionId,
-  isImageItem,
-  isVideoItem,
-  isMediaItem,
-  isCollectionItem,
-  type ImageTimelineItem,
-  type VideoTimelineItem,
-  type CollectionTimelineItem,
-  type TimelineItemId,
-  type CollectionId,
-  type TimelineCollection,
-  type TimelineItem,
-} from "./media-strip.types";
-
-import {
-  validateTimelineItemTiming,
-  validateCollectionTimelineItem,
-  validateVideoTimelineItem,
-  validateTimelineItem,
-  validateMediaItemStrings,
-  createImageTimelineItem,
-  createCollectionTimelineItem,
-  createVideoTimelineItem,
-  updateImageTimelineItem,
-  updateCollectionTimelineItem,
-  updateVideoTimelineItem,
-  validateTimelineCollection,
-  wouldCreateCollectionCycle,
-} from "./media-strip.validation";
-import { validateProjectTimeline } from "./media-strip.project-validation";
-
-import {
-  getTimelineItemEndTimeSeconds,
-  getVideoVisibleDurationSeconds,
-  getItemWidth,
-  formatDuration,
-  areEqual,
-  MIN_ITEM_WIDTH_PX,
-  DRAG_ACTIVATION_THRESHOLDS_PX,
-} from "./media-strip.utils";
-import {
-  encodeDndTarget,
-  decodeDndTarget,
-  detectCollision,
-} from "./media-strip.dnd";
-import {
-  syncCollectionItemCounts,
-  applyTimelineItemCommand,
-} from "./media-strip.collection-ops";
-
-describe("Timeline Items Branding", () => {
-  test("validates timeline item and collection IDs", () => {
-    const validItem = parseTimelineItemId("item-1");
-    const validCol = parseCollectionId("collection-1");
-    const invalidItem = parseTimelineItemId("");
-    const invalidCol = parseCollectionId("");
-    const whitespaceItem = parseTimelineItemId("   ");
-    const whitespaceCol = parseCollectionId("   ");
-
-    expect(validItem).toEqual({ ok: true, value: "item-1" as TimelineItemId });
-    expect(validCol).toEqual({ ok: true, value: "collection-1" as CollectionId });
-    expect(invalidItem).toEqual({ ok: false, error: "empty-id" });
-    expect(invalidCol).toEqual({ ok: false, error: "empty-id" });
-    expect(whitespaceItem).toEqual({ ok: false, error: "empty-id" });
-    expect(whitespaceCol).toEqual({ ok: false, error: "empty-id" });
-  });
-});
-
-describe("Timeline Items Type Guards", () => {
-  const imageItem: ImageTimelineItem = {
-    id: "image-1" as TimelineItemId,
-    name: "Test Image",
-    kind: "image",
-    src: "http://example.com/image.jpg",
-    startTimeSeconds: 0,
-    durationSeconds: 10,
-  };
-
-  const videoItem: VideoTimelineItem = {
-    id: "video-1" as TimelineItemId,
-    name: "Test Video",
-    kind: "video",
-    src: "http://example.com/video.mp4",
-    startTimeSeconds: 10,
-    durationSeconds: 20,
-    sourceDurationSeconds: 30,
-    trimInSeconds: 5,
-    trimOutSeconds: 5,
-  };
-
-  const collectionItem: CollectionTimelineItem = {
-    id: "collection-1" as TimelineItemId,
-    name: "Test Collection",
-    kind: "collection",
-    collectionId: "collection-col-1" as CollectionId,
-    itemCount: 3,
-    startTimeSeconds: 5,
-    durationSeconds: 15,
-  };
-
-  test("correctly identifies image items", () => {
-    expect(isImageItem(imageItem)).toBe(true);
-    expect(isImageItem(videoItem)).toBe(false);
-    expect(isImageItem(collectionItem)).toBe(false);
-  });
-
-  test("correctly identifies video items", () => {
-    expect(isVideoItem(imageItem)).toBe(false);
-    expect(isVideoItem(videoItem)).toBe(true);
-    expect(isVideoItem(collectionItem)).toBe(false);
-  });
-
-  test("correctly identifies media items", () => {
-    expect(isMediaItem(imageItem)).toBe(true);
-    expect(isMediaItem(videoItem)).toBe(true);
-    expect(isMediaItem(collectionItem)).toBe(false);
-  });
-
-  test("correctly identifies collection items", () => {
-    expect(isCollectionItem(imageItem)).toBe(false);
-    expect(isCollectionItem(videoItem)).toBe(false);
-    expect(isCollectionItem(collectionItem)).toBe(true);
-  });
-});
-
-describe("Timeline Items Derived Helpers", () => {
-  test("calculates end time correctly", () => {
-    const item = {
-      id: "item-1" as TimelineItemId,
-      name: "Item",
-      startTimeSeconds: 5.5,
-      durationSeconds: 10.2,
-    };
-    expect(getTimelineItemEndTimeSeconds(item as any)).toBeCloseTo(15.7);
-  });
-
-  test("calculates video visible duration correctly", () => {
-    const videoData = {
-      sourceDurationSeconds: 120.5,
-      trimInSeconds: 10.2,
-      trimOutSeconds: 15.3,
-    };
-    expect(getVideoVisibleDurationSeconds(videoData)).toBeCloseTo(95.0);
-  });
-});
-
-describe("Timeline Items Timing Validation", () => {
-  test("validates correct timing", () => {
-    const item = {
-      id: "item-1" as TimelineItemId,
-      name: "Item",
-      startTimeSeconds: 0,
-      durationSeconds: 10,
-    };
-    expect(validateTimelineItemTiming(item as any)).toEqual({ valid: true });
-  });
-
-  test("fails timing validation for negative start time", () => {
-    const item = {
-      id: "item-1" as TimelineItemId,
-      name: "Item",
-      startTimeSeconds: -1,
-      durationSeconds: 10,
-    };
-    expect(validateTimelineItemTiming(item as any)).toEqual({
-      valid: false,
-      reason: "negative-start-time",
-    });
-  });
-
-  test("fails timing validation for negative duration", () => {
-    const item = {
-      id: "item-1" as TimelineItemId,
-      name: "Item",
-      startTimeSeconds: 0,
-      durationSeconds: -5,
-    };
-    expect(validateTimelineItemTiming(item as any)).toEqual({
-      valid: false,
-      reason: "negative-duration",
-    });
-  });
-
-  test("fails timing validation for non-finite values", () => {
-    const itemStartInfinity = {
-      id: "item-1" as TimelineItemId,
-      name: "Item",
-      startTimeSeconds: Infinity,
-      durationSeconds: 10,
-    };
-    expect(validateTimelineItemTiming(itemStartInfinity as any)).toEqual({
-      valid: false,
-      reason: "non-finite-start-time",
-    });
-
-    const itemDurationNaN = {
-      id: "item-1" as TimelineItemId,
-      name: "Item",
-      startTimeSeconds: 0,
-      durationSeconds: NaN,
-    };
-    expect(validateTimelineItemTiming(itemDurationNaN as any)).toEqual({
-      valid: false,
-      reason: "non-finite-duration",
-    });
-  });
-
-  test("fails timing validation for empty or invalid name", () => {
-    const itemEmptyName = {
-      id: "item-1" as TimelineItemId,
-      name: "",
-      startTimeSeconds: 0,
-      durationSeconds: 10,
-    };
-    expect(validateTimelineItemTiming(itemEmptyName as any)).toEqual({
-      valid: false,
-      reason: "empty-name",
-    });
-
-    const itemWhitespaceName = {
-      id: "item-1" as TimelineItemId,
-      name: "   ",
-      startTimeSeconds: 0,
-      durationSeconds: 10,
-    };
-    expect(validateTimelineItemTiming(itemWhitespaceName as any)).toEqual({
-      valid: false,
-      reason: "empty-name",
-    });
-  });
-
-  test("tolerates tiny negative times within epsilon", () => {
-    const itemTinyNegativeStart = {
-      id: "item-1" as TimelineItemId,
-      name: "Item",
-      startTimeSeconds: -0.0000001,
-      durationSeconds: 10,
-    };
-    expect(validateTimelineItemTiming(itemTinyNegativeStart as any)).toEqual({
-      valid: true,
-    });
-  });
-});
-
-describe("Media Items String Validation", () => {
-  test("validates correct string values", () => {
-    const item = {
-      src: "https://example.com/video.mp4",
-      posterSrcs: ["https://example.com/frame1.jpg", "https://example.com/frame2.jpg"],
-    };
-    expect(validateMediaItemStrings(item as any)).toEqual({ valid: true });
-  });
-
-  test("fails for invalid src type", () => {
-    const item = {
-      src: 123,
-    };
-    expect(validateMediaItemStrings(item as any)).toEqual({
-      valid: false,
-      reason: "invalid-src",
-    });
-  });
-
-  test("fails for empty or whitespace src", () => {
-    const itemEmpty = { src: "" };
-    expect(validateMediaItemStrings(itemEmpty as any)).toEqual({
-      valid: false,
-      reason: "invalid-src",
-    });
-
-    const itemWhitespace = { src: "   " };
-    expect(validateMediaItemStrings(itemWhitespace as any)).toEqual({
-      valid: false,
-      reason: "invalid-src",
-    });
-  });
-
-  test("fails for invalid posterSrcs array or elements", () => {
-    const itemNotArray = {
-      src: "https://example.com/video.mp4",
-      posterSrcs: "not-an-array",
-    };
-    expect(validateMediaItemStrings(itemNotArray as any)).toEqual({
-      valid: false,
-      reason: "invalid-poster-srcs",
-    });
-
-    const itemInvalidElements = {
-      src: "https://example.com/video.mp4",
-      posterSrcs: ["https://example.com/frame1.jpg", 123],
-    };
-    expect(validateMediaItemStrings(itemInvalidElements as any)).toEqual({
-      valid: false,
-      reason: "invalid-poster-srcs",
-    });
-
-    const itemEmptyElement = {
-      src: "https://example.com/video.mp4",
-      posterSrcs: ["https://example.com/frame1.jpg", ""],
-    };
-    expect(validateMediaItemStrings(itemEmptyElement as any)).toEqual({
-      valid: false,
-      reason: "invalid-poster-srcs",
-    });
-  });
-});
-
-describe("Collection Timing and Integrity Validation", () => {
-  const baseCollection = {
-    id: "col-1" as TimelineItemId,
-    name: "Col",
-    kind: "collection" as const,
-    collectionId: "c-1" as CollectionId,
-    startTimeSeconds: 0,
-    durationSeconds: 10,
-    itemCount: 5,
-  };
-
-  test("validates correct collection", () => {
-    expect(validateCollectionTimelineItem(baseCollection)).toEqual({ valid: true });
-  });
-
-  test("fails for invalid timing in collection", () => {
-    expect(
-      validateCollectionTimelineItem({
-        ...baseCollection,
-        startTimeSeconds: -1,
-      })
-    ).toEqual({ valid: false, reason: "negative-start-time" });
-  });
-
-  test("fails for empty collectionId", () => {
-    expect(
-      validateCollectionTimelineItem({
-        ...baseCollection,
-        collectionId: "" as CollectionId,
-      })
-    ).toEqual({ valid: false, reason: "empty-id" });
-  });
-
-  test("fails for invalid item counts", () => {
-    expect(
-      validateCollectionTimelineItem({
-        ...baseCollection,
-        itemCount: -5,
-      })
-    ).toEqual({ valid: false, reason: "negative-item-count" });
-
-    expect(
-      validateCollectionTimelineItem({
-        ...baseCollection,
-        itemCount: 2.5,
-      })
-    ).toEqual({ valid: false, reason: "non-integer-item-count" });
-
-    expect(
-      validateCollectionTimelineItem({
-        ...baseCollection,
-        itemCount: Infinity,
-      })
-    ).toEqual({ valid: false, reason: "non-finite-item-count" });
-  });
-});
-
-describe("Video Timing and Integrity Validation", () => {
-  const baseVideo: VideoTimelineItem = {
-    id: "vid-1" as TimelineItemId,
-    name: "Vid",
-    kind: "video",
-    src: "vid.mp4",
-    startTimeSeconds: 0,
-    durationSeconds: 10,
-    sourceDurationSeconds: 20,
-    trimInSeconds: 5,
-    trimOutSeconds: 5,
-  };
-
-  test("validates correct video", () => {
-    expect(validateVideoTimelineItem(baseVideo)).toEqual({ valid: true });
-  });
-
-  test("fails for negative trim/duration values", () => {
-    expect(
-      validateVideoTimelineItem({
-        ...baseVideo,
-        sourceDurationSeconds: -20,
-        durationSeconds: 0,
-        trimInSeconds: 0,
-        trimOutSeconds: 20,
-      })
-    ).toEqual({ valid: false, reason: "negative-source-duration" });
-
-    expect(
-      validateVideoTimelineItem({
-        ...baseVideo,
-        trimInSeconds: -2,
-        durationSeconds: 12,
-      })
-    ).toEqual({ valid: false, reason: "negative-trim-in" });
-  });
-
-  test("fails if trim exceeds source duration", () => {
-    expect(
-      validateVideoTimelineItem({
-        ...baseVideo,
-        trimInSeconds: 15,
-        trimOutSeconds: 10,
-        durationSeconds: 0,
-      })
-    ).toEqual({ valid: false, reason: "trim-exceeds-source" });
-  });
-
-  test("fails if duration mismatch occurs", () => {
-    expect(
-      validateVideoTimelineItem({
-        ...baseVideo,
-        durationSeconds: 8, // Should be 10 (20 - 5 - 5)
-      })
-    ).toEqual({
-      valid: false,
-      reason: "duration-mismatch",
-      expectedDurationSeconds: 10,
-    });
-  });
-});
-
-describe("Timeline Items Smart Constructors", () => {
-  test("creates image item successfully", () => {
-    const result = createImageTimelineItem({
-      id: "img-1" as TimelineItemId,
-      name: "Img",
-      src: "img.png",
-      startTimeSeconds: -0.0000001, // normalized to 0
-      durationSeconds: 15,
-    });
-
-    expect(result.ok).toBe(true);
-    if (result.ok) {
-      expect(result.value.startTimeSeconds).toBe(0);
-      expect(result.value.kind).toBe("image");
-    }
-  });
-
-  test("fails creating image item on validation error", () => {
-    const result = createImageTimelineItem({
-      id: "img-1" as TimelineItemId,
-      name: "Img",
-      src: "img.png",
-      startTimeSeconds: -5,
-      durationSeconds: 15,
-    });
-
-    expect(result.ok).toBe(false);
-    if (!result.ok) {
-      expect(result.error.reason).toBe("negative-start-time");
-    }
-  });
-
-  test("creates collection item successfully", () => {
-    const result = createCollectionTimelineItem({
-      id: "col-1" as TimelineItemId,
-      name: "Col",
-      collectionId: "c-1" as CollectionId,
-      startTimeSeconds: 0,
-      durationSeconds: 10,
-      itemCount: 4,
-    });
-
-    expect(result.ok).toBe(true);
-    if (result.ok) {
-      expect(result.value.itemCount).toBe(4);
-      expect(result.value.kind).toBe("collection");
-    }
-  });
-
-  test("creates video item and automatically calculates duration", () => {
-    const result = createVideoTimelineItem({
-      id: "vid-1" as TimelineItemId,
-      name: "Vid",
-      src: "vid.mp4",
-      startTimeSeconds: 2,
-      sourceDurationSeconds: 50,
-      trimInSeconds: 10.5,
-      trimOutSeconds: 9.5,
-    });
-
-    expect(result.ok).toBe(true);
-    if (result.ok) {
-      expect(result.value.durationSeconds).toBe(30); // 50 - 10.5 - 9.5
-      expect(result.value.kind).toBe("video");
-    }
-  });
-});
-
-describe("Timeline Items Update Helpers", () => {
-  test("updates image item and normalizes timings", () => {
-    const image = createImageTimelineItem({
-      id: "img-1" as TimelineItemId,
-      name: "Img",
-      src: "img.png",
-      startTimeSeconds: 0,
-      durationSeconds: 10,
-    });
-
-    expect(image.ok).toBe(true);
-    if (image.ok) {
-      const updated = updateImageTimelineItem(image.value, {
-        startTimeSeconds: 5,
-        durationSeconds: 20,
-      });
-      expect(updated.ok).toBe(true);
-      if (updated.ok) {
-        expect(updated.value.startTimeSeconds).toBe(5);
-        expect(updated.value.durationSeconds).toBe(20);
-        expect(updated.value.name).toBe("Img");
-      }
-    }
-  });
-
-  test("updates video item and recalculates duration", () => {
-    const video = createVideoTimelineItem({
-      id: "vid-1" as TimelineItemId,
-      name: "Vid",
-      src: "vid.mp4",
-      startTimeSeconds: 0,
-      sourceDurationSeconds: 100,
-      trimInSeconds: 10,
-      trimOutSeconds: 10,
-    });
-
-    expect(video.ok).toBe(true);
-    if (video.ok) {
-      const updated = updateVideoTimelineItem(video.value, {
-        trimInSeconds: 20,
-        trimOutSeconds: 30,
-      });
-      expect(updated.ok).toBe(true);
-      if (updated.ok) {
-        expect(updated.value.durationSeconds).toBe(50); // 100 - 20 - 30
-      }
-    }
-  });
-});
-
-describe("MediaStrip formatDuration helper", () => {
-  test("formats short durations under a minute", () => {
-    expect(formatDuration(0)).toBe("00:00");
-    expect(formatDuration(5)).toBe("00:05");
-    expect(formatDuration(59.9)).toBe("00:59");
-  });
-
-  test("formats durations under an hour", () => {
-    expect(formatDuration(60)).toBe("01:00");
-    expect(formatDuration(75)).toBe("01:15");
-    expect(formatDuration(3599)).toBe("59:59");
-  });
-
-  test("formats durations greater than or equal to an hour", () => {
-    expect(formatDuration(3600)).toBe("1:00:00");
-    expect(formatDuration(3665)).toBe("1:01:05");
-    expect(formatDuration(7200)).toBe("2:00:00");
-    expect(formatDuration(90065)).toBe("25:01:05");
-  });
-});
-
-describe("MediaStrip getItemWidth helper", () => {
-  test("clamps width to minimum size", () => {
-    const item = { durationSeconds: 1 } as any;
-    // 1s * 32px/s = 32px, clamped to MIN_ITEM_WIDTH_PX (96px)
-    expect(getItemWidth(item, 32)).toBe(MIN_ITEM_WIDTH_PX);
-  });
-
-  test("does not clamp width to a maximum size", () => {
-    const item = { durationSeconds: 20 } as any;
-    // 20s * 32px/s = 640px
-    expect(getItemWidth(item, 32)).toBe(640);
-  });
-
-  test("returns linearly scaled width within boundaries", () => {
-    const item = { durationSeconds: 5 } as any;
-    // 5s * 32px/s = 160px
-    expect(getItemWidth(item, 32)).toBe(160);
-  });
-
-  test("handles alternative pxPerSecond scales", () => {
-    const item = { durationSeconds: 5 } as any;
-    // 5s * 20px/s = 100px
-    expect(getItemWidth(item, 20)).toBe(100);
-  });
-});
-
-describe("MediaStripItemButton areEqual custom comparator", () => {
-  const itemA = { id: "item-a" as TimelineItemId, name: "Item A" } as any;
-  const itemB = { id: "item-b" as TimelineItemId, name: "Item B" } as any;
-  const index = 0;
-
-  test("returns true for identical items and matching styles", () => {
-    const prev = {
-      item: itemA,
-      index,
-      isKeyboardReordering: false,
-      style: {
-        width: "100px",
-        left: "10px",
-        top: 4,
-        height: "calc(100% - 8px)",
-      },
-    };
-    const next = {
-      item: itemA,
-      index,
-      isKeyboardReordering: false,
-      style: {
-        width: "100px",
-        left: "10px",
-        top: 4,
-        height: "calc(100% - 8px)",
-      },
-    };
-    expect(areEqual(prev, next)).toBe(true);
-  });
-
-  test("returns false if items change", () => {
-    const prev = {
-      item: itemA,
-      index,
-      isKeyboardReordering: false,
-      style: {
-        width: "100px",
-        left: "10px",
-        top: 4,
-        height: "calc(100% - 8px)",
-      },
-    };
-    const next = {
-      item: itemB,
-      index,
-      isKeyboardReordering: false,
-      style: {
-        width: "100px",
-        left: "10px",
-        top: 4,
-        height: "calc(100% - 8px)",
-      },
-    };
-    expect(areEqual(prev, next)).toBe(false);
-  });
-
-  test("returns false if style width changes", () => {
-    const prev = {
-      item: itemA,
-      index,
-      isKeyboardReordering: false,
-      style: {
-        width: "100px",
-        left: "10px",
-        top: 4,
-        height: "calc(100% - 8px)",
-      },
-    };
-    const next = {
-      item: itemA,
-      index,
-      isKeyboardReordering: false,
-      style: {
-        width: "120px",
-        left: "10px",
-        top: 4,
-        height: "calc(100% - 8px)",
-      },
-    };
-    expect(areEqual(prev, next)).toBe(false);
-  });
-
-  test("returns false if style left changes", () => {
-    const prev = {
-      item: itemA,
-      index,
-      isKeyboardReordering: false,
-      style: {
-        width: "100px",
-        left: "10px",
-        top: 4,
-        height: "calc(100% - 8px)",
-      },
-    };
-    const next = {
-      item: itemA,
-      index,
-      isKeyboardReordering: false,
-      style: {
-        width: "100px",
-        left: "20px",
-        top: 4,
-        height: "calc(100% - 8px)",
-      },
-    };
-    expect(areEqual(prev, next)).toBe(false);
-  });
-
-  test("returns false if style top changes", () => {
-    const prev = {
-      item: itemA,
-      index,
-      isKeyboardReordering: false,
-      style: {
-        width: "100px",
-        left: "10px",
-        top: 4,
-        height: "calc(100% - 8px)",
-      },
-    };
-    const next = {
-      item: itemA,
-      index,
-      isKeyboardReordering: false,
-      style: {
-        width: "100px",
-        left: "10px",
-        top: 6,
-        height: "calc(100% - 8px)",
-      },
-    };
-    expect(areEqual(prev, next)).toBe(false);
-  });
-
-  test("returns false if style height changes", () => {
-    const prev = {
-      item: itemA,
-      index,
-      isKeyboardReordering: false,
-      style: {
-        width: "100px",
-        left: "10px",
-        top: 4,
-        height: "calc(100% - 8px)",
-      },
-    };
-    const next = {
-      item: itemA,
-      index,
-      isKeyboardReordering: false,
-      style: {
-        width: "100px",
-        left: "10px",
-        top: 4,
-        height: "calc(100% - 12px)",
-      },
-    };
-    expect(areEqual(prev, next)).toBe(false);
-  });
-
-  test("returns false if isKeyboardReordering changes", () => {
-    const prev = {
-      item: itemA,
-      index,
-      isKeyboardReordering: false,
-      style: {
-        width: "100px",
-        left: "10px",
-        top: 4,
-        height: "calc(100% - 8px)",
-      },
-    };
-    const next = {
-      item: itemA,
-      index,
-      isKeyboardReordering: true,
-      style: {
-        width: "100px",
-        left: "10px",
-        top: 4,
-        height: "calc(100% - 8px)",
-      },
-    };
-    expect(areEqual(prev, next)).toBe(false);
-  });
-
-  test("handles missing or undefined style objects safely", () => {
-    const prev = {
-      item: itemA,
-      index,
-      isKeyboardReordering: false,
-    } as any;
-    const next = {
-      item: itemA,
-      index,
-      isKeyboardReordering: false,
-      style: {
-        width: "100px",
-        left: "10px",
-        top: 4,
-        height: "calc(100% - 8px)",
-      },
-    } as any;
-    expect(areEqual(prev, next)).toBe(false);
-    expect(areEqual(next, prev)).toBe(false);
-    expect(areEqual(prev, prev)).toBe(true);
-  });
-});
-
-describe("validateTimelineItem dynamic dispatcher", () => {
-  test("correctly dispatches validation to image timeline items", () => {
-    const image = {
-      id: "img-1" as TimelineItemId,
-      kind: "image",
-      name: "Img",
-      src: "img.jpg",
-      startTimeSeconds: 0,
-      durationSeconds: 10,
-    } as any;
-    expect(validateTimelineItem(image)).toEqual({ valid: true });
-
-    const invalidImage = { ...image, startTimeSeconds: -5 };
-    expect(validateTimelineItem(invalidImage)).toEqual({
-      valid: false,
-      reason: "negative-start-time",
-    });
-  });
-
-  test("correctly dispatches validation to video timeline items", () => {
-    const video = {
-      id: "vid-1" as TimelineItemId,
-      kind: "video",
-      name: "Vid",
-      src: "vid.mp4",
-      startTimeSeconds: 0,
-      durationSeconds: 10,
-      sourceDurationSeconds: 100,
-      trimInSeconds: 10,
-      trimOutSeconds: 80,
-    } as any;
-    expect(validateTimelineItem(video)).toEqual({ valid: true });
-
-    const invalidVideo = { ...video, trimInSeconds: -5 };
-    expect(validateTimelineItem(invalidVideo)).toEqual({
-      valid: false,
-      reason: "negative-trim-in",
-    });
-  });
-
-  test("correctly dispatches validation to collection timeline items", () => {
-    const collection = {
-      id: "col-1" as TimelineItemId,
-      kind: "collection",
-      name: "Col",
-      startTimeSeconds: 0,
-      durationSeconds: 10,
-      collectionId: "col-id" as CollectionId,
-      itemCount: 5,
-    } as any;
-    expect(validateTimelineItem(collection)).toEqual({ valid: true });
-
-    const invalidCollection = { ...collection, itemCount: -1 };
-    expect(validateTimelineItem(invalidCollection)).toEqual({
-      valid: false,
-      reason: "negative-item-count",
-    });
-  });
-});
-
-describe("Drag and Scroll Activation Thresholds", () => {
-  test("asserts scroll threshold is strictly less than board threshold", () => {
-    // This pins the UX constraint: diagonal/small drags below 5px prioritize
-    // horizontal scroll area scrolling over board-level item reordering.
-    expect(DRAG_ACTIVATION_THRESHOLDS_PX.scroll).toBeLessThan(
-      DRAG_ACTIVATION_THRESHOLDS_PX.board
-    );
-  });
-});
-
-describe("Video Constructor Failure Ordering & Trim Validation", () => {
-  test("createVideoTimelineItem returns trim-exceeds-source when derived duration would be negative", () => {
-    const result = createVideoTimelineItem({
-      id: "vid-1",
-      name: "Vid",
-      src: "vid.mp4",
-      startTimeSeconds: 0,
-      sourceDurationSeconds: 10,
-      trimInSeconds: 8,
-      trimOutSeconds: 8,
-    });
-
-    expect(result.ok).toBe(false);
-    if (!result.ok) {
-      expect(result.error.reason).toBe("trim-exceeds-source");
-    }
-  });
-
-  test("updateVideoTimelineItem rejects invalid trim updates", () => {
-    const video = {
-      id: "vid-1" as TimelineItemId,
-      name: "Video",
-      kind: "video" as const,
-      src: "vid.mp4",
-      startTimeSeconds: 0,
-      sourceDurationSeconds: 10,
-      trimInSeconds: 2,
-      trimOutSeconds: 2,
-      durationSeconds: 6,
-    };
-
-    // trimInSeconds: -1
-    const res1 = updateVideoTimelineItem(video, { trimInSeconds: -1 });
-    expect(res1.ok).toBe(false);
-
-    // trimOutSeconds: -1
-    const res2 = updateVideoTimelineItem(video, { trimOutSeconds: -1 });
-    expect(res2.ok).toBe(false);
-
-    // trimInSeconds + trimOutSeconds > sourceDurationSeconds
-    const res3 = updateVideoTimelineItem(video, { trimInSeconds: 6, trimOutSeconds: 6 });
-    expect(res3.ok).toBe(false);
-    if (!res3.ok) {
-      expect(res3.error.reason).toBe("trim-exceeds-source");
-    }
-
-    // sourceDurationSeconds: NaN
-    const res4 = updateVideoTimelineItem(video, { sourceDurationSeconds: NaN });
-    expect(res4.ok).toBe(false);
-  });
-});
-
-describe("Tiny Negative Normalization for all constructors", () => {
-  test("normalizes tiny negative startTimeSeconds for collection timeline items", () => {
-    const result = createCollectionTimelineItem({
-      id: "col-1",
-      name: "Collection",
-      collectionId: "collection-col",
-      itemCount: 5,
-      startTimeSeconds: -0.0000001,
-      durationSeconds: 10,
-    });
-
-    expect(result.ok).toBe(true);
-    if (result.ok) {
-      expect(result.value.startTimeSeconds).toBe(0);
-    }
-  });
-
-  test("normalizes tiny negative trimInSeconds for video timeline items", () => {
-    const result = createVideoTimelineItem({
-      id: "vid-1",
-      name: "Video",
-      src: "vid.mp4",
-      startTimeSeconds: 0,
-      sourceDurationSeconds: 10,
-      trimInSeconds: -0.0000001,
-      trimOutSeconds: 2,
-    });
-
-    expect(result.ok).toBe(true);
-    if (result.ok) {
-      expect(result.value.trimInSeconds).toBe(0);
-    }
-  });
-});
-
-describe("updateCollectionTimelineItem failure cases", () => {
-  const collection = {
-    id: "col-1" as TimelineItemId,
-    name: "Collection",
-    kind: "collection" as const,
-    collectionId: "collection-col" as CollectionId,
-    itemCount: 5,
-    startTimeSeconds: 0,
-    durationSeconds: 10,
-  };
-
-  test("rejects negative itemCount", () => {
-    const result = updateCollectionTimelineItem(collection, { itemCount: -1 });
-    expect(result.ok).toBe(false);
-    if (!result.ok) {
-      expect(result.error.reason).toBe("negative-item-count");
-    }
-  });
-
-  test("rejects non-integer itemCount", () => {
-    const result = updateCollectionTimelineItem(collection, { itemCount: 1.5 });
-    expect(result.ok).toBe(false);
-    if (!result.ok) {
-      expect(result.error.reason).toBe("non-integer-item-count");
-    }
-  });
-
-  test("rejects empty collectionId", () => {
-    const result = updateCollectionTimelineItem(collection, { collectionId: "" as any });
-    expect(result.ok).toBe(false);
-    if (!result.ok) {
-      expect(result.error.reason).toBe("empty-id");
-    }
-  });
-
-  test("rejects negative durationSeconds", () => {
-    const result = updateCollectionTimelineItem(collection, { durationSeconds: -1 });
-    expect(result.ok).toBe(false);
-    if (!result.ok) {
-      expect(result.error.reason).toBe("negative-duration");
-    }
-  });
-});
-
-describe("getItemWidth defensive behavior and formatDuration edge cases", () => {
-  test("getItemWidth returns MIN_ITEM_WIDTH_PX on invalid inputs", () => {
-    expect(getItemWidth({ durationSeconds: NaN }, 32)).toBe(MIN_ITEM_WIDTH_PX);
-    expect(getItemWidth({ durationSeconds: Infinity }, 0)).toBe(MIN_ITEM_WIDTH_PX);
-    expect(getItemWidth({ durationSeconds: 5 }, NaN)).toBe(MIN_ITEM_WIDTH_PX);
-    expect(getItemWidth({ durationSeconds: 5 }, -32)).toBe(MIN_ITEM_WIDTH_PX);
-  });
-
-  test("formatDuration handles non-finite values and negative inputs", () => {
-    expect(formatDuration(-1)).toBe("00:00");
-    expect(formatDuration(NaN)).toBe("00:00");
-    expect(formatDuration(Infinity)).toBe("00:00");
-  });
-});
-
-describe("areEqual additional comparator tests", () => {
-  const itemA = { id: "item-a" as TimelineItemId, name: "Item A" } as any;
-
-  const baseProps = {
-    item: itemA,
-    index: 0,
-    isKeyboardReordering: false,
-    thumbnailVariant: "sequence" as const,
-    collectionId: "strip-1" as CollectionId,
-    onMoveItem: () => { },
-    style: {
-      width: "100px",
-      left: "10px",
-      top: 4,
-      height: "calc(100% - 8px)",
-    },
-  };
-
-  test("returns false when thumbnailVariant changes", () => {
-    const prev = { ...baseProps, thumbnailVariant: "sequence" as const };
-    const next = { ...baseProps, thumbnailVariant: "single" as const };
-    expect(areEqual(prev, next)).toBe(false);
-  });
-
-  test("returns false when collectionId changes", () => {
-    const prev = { ...baseProps, collectionId: "strip-1" as CollectionId };
-    const next = { ...baseProps, collectionId: "strip-2" as CollectionId };
-    expect(areEqual(prev, next)).toBe(false);
-  });
-
-  test("returns false when onMoveItem reference changes", () => {
-    const prev = { ...baseProps, onMoveItem: () => { } };
-    const next = { ...baseProps, onMoveItem: () => { } };
-    expect(areEqual(prev, next)).toBe(false);
-  });
-
-  test("returns false when index changes", () => {
-    const prev = { ...baseProps, index: 0 };
-    const next = { ...baseProps, index: 1 };
-    expect(areEqual(prev, next)).toBe(false);
-  });
-});
-
-describe("Deeply Nested Collections Dnd Target encoding/decoding", () => {
-  test("correctly encodes and decodes DndTargets", () => {
-    const itemTarget = { type: "item" as const, itemId: "item-1" as TimelineItemId };
-    const containerTarget = { type: "collection-container" as const, collectionId: "col-1" as CollectionId };
-    const nestTarget = { type: "collection-nest-target" as const, collectionId: "col-2" as CollectionId };
-
-    const encodedItem = encodeDndTarget(itemTarget);
-    const encodedContainer = encodeDndTarget(containerTarget);
-    const encodedNest = encodeDndTarget(nestTarget);
-
-    expect(encodedItem).toBe("item:item-1");
-    expect(encodedContainer).toBe("container:col-1");
-    expect(encodedNest).toBe("nest:col-2");
-
-    expect(decodeDndTarget(encodedItem)).toEqual(itemTarget);
-    expect(decodeDndTarget(encodedContainer)).toEqual(containerTarget);
-    expect(decodeDndTarget(encodedNest)).toEqual(nestTarget);
-    expect(decodeDndTarget("invalid-prefix:something")).toBeNull();
-  });
-});
-
-describe("Cycle Detection for Nested Collections", () => {
-  const colA: TimelineCollection = {
-    id: "col-a" as CollectionId,
-    name: "Collection A",
-    items: [
-      {
-        id: "item-a" as TimelineItemId,
-        name: "Pointer to B",
-        kind: "collection",
-        collectionId: "col-b" as CollectionId,
-        itemCount: 0,
-        startTimeSeconds: 0,
-        durationSeconds: 10,
-      },
-    ],
-  };
-
-  const colB: TimelineCollection = {
-    id: "col-b" as CollectionId,
-    name: "Collection B",
-    items: [],
-  };
-
-  const collections = new Map<CollectionId, TimelineCollection>([
-    ["col-a" as CollectionId, colA],
-    ["col-b" as CollectionId, colB],
-  ]);
-
-  test("detects when nesting creates a cycle", () => {
-    expect(
-      wouldCreateCollectionCycle({
-        movingCollectionId: "col-a" as CollectionId,
-        targetCollectionId: "col-b" as CollectionId,
-        collectionsById: collections,
-      })
-    ).toBe(true);
-
-    expect(
-      wouldCreateCollectionCycle({
-        movingCollectionId: "col-b" as CollectionId,
-        targetCollectionId: "col-a" as CollectionId,
-        collectionsById: collections,
-      })
-    ).toBe(false);
-  });
-});
-
-describe("Collection and Project Validation", () => {
-  test("validates a collection correctly", () => {
-    const validCol: TimelineCollection = {
-      id: "col-1" as CollectionId,
-      name: "Valid Collection",
-      items: [
-        {
-          id: "item-1" as TimelineItemId,
-          name: "Image Item",
-          kind: "image",
-          src: "image.jpg",
-          startTimeSeconds: 0,
-          durationSeconds: 10,
-        },
-      ],
-    };
-    expect(validateTimelineCollection(validCol).valid).toBe(true);
-
-    const invalidCol: TimelineCollection = {
-      id: "" as CollectionId,
-      name: "",
-      items: [],
-    };
-    expect(validateTimelineCollection(invalidCol).valid).toBe(false);
-  });
-
-  test("validates a project timeline correctly", () => {
-    const colA: TimelineCollection = {
-      id: "col-a" as CollectionId,
-      name: "A",
-      items: [
-        {
-          id: "item-b" as TimelineItemId,
-          name: "B Card",
-          kind: "collection",
-          collectionId: "col-b" as CollectionId,
-          itemCount: 0,
-          startTimeSeconds: 0,
-          durationSeconds: 5,
-        },
-      ],
-    };
-
-    const colB: TimelineCollection = {
-      id: "col-b" as CollectionId,
-      name: "B",
-      items: [
-        {
-          id: "item-a" as TimelineItemId,
-          name: "A Card (Cycle)",
-          kind: "collection",
-          collectionId: "col-a" as CollectionId,
-          itemCount: 0,
-          startTimeSeconds: 0,
-          durationSeconds: 5,
-        },
-      ],
-    };
-
-    const collections = new Map<CollectionId, TimelineCollection>([
-      ["col-a" as CollectionId, colA],
-      ["col-b" as CollectionId, colB],
-    ]);
-
-    const result = validateProjectTimeline({
-      collectionsById: collections,
-      rootCollectionIds: ["col-a" as CollectionId],
-    });
-
-    expect(result.valid).toBe(false);
-    if (!result.valid) {
-      expect(result.reason).toBe("collection-cycle");
-    }
-  });
-
-  test("returns orphanedCollectionIds for unreachable collections from roots", () => {
-    const colA: TimelineCollection = {
-      id: "col-a" as CollectionId,
-      name: "A",
-      items: [],
-    };
-
-    const colOrphan: TimelineCollection = {
-      id: "col-orphan" as CollectionId,
-      name: "Orphaned",
-      items: [],
-    };
-
-    const collections = new Map<CollectionId, TimelineCollection>([
-      ["col-a" as CollectionId, colA],
-      ["col-orphan" as CollectionId, colOrphan],
-    ]);
-
-    const result = validateProjectTimeline({
-      collectionsById: collections,
-      rootCollectionIds: ["col-a" as CollectionId],
-    });
-
-    expect(result.valid).toBe(true);
-    if (result.valid) {
-      expect(result.orphanedCollectionIds).toEqual(["col-orphan"]);
-    }
-  });
-});
-
-describe("Collection Move and Count synchronization", () => {
-  test("moves item and synchronizes counts correctly", () => {
-    const item1 = {
-      id: "item-1" as TimelineItemId,
-      name: "Image 1",
-      kind: "image" as const,
-      src: "image1.jpg",
-      startTimeSeconds: 0,
-      durationSeconds: 10,
-    };
-    const colA: TimelineCollection = {
-      id: "col-a" as CollectionId,
-      name: "Col A",
-      items: [item1],
-    };
-    const colB: TimelineCollection = {
-      id: "col-b" as CollectionId,
-      name: "Col B",
-      items: [],
-    };
-
-    const collections = new Map<CollectionId, TimelineCollection>([
-      ["col-a" as CollectionId, colA],
-      ["col-b" as CollectionId, colB],
-    ]);
-
-    const result = applyTimelineItemCommand({
-      collectionsById: collections,
-      command: {
-        type: "move",
-        itemId: "item-1" as TimelineItemId,
-        fromCollectionId: "col-a" as CollectionId,
-        toCollectionId: "col-b" as CollectionId,
-        toIndex: 0,
-      },
-    });
-
-    expect(result.get("col-a" as CollectionId)?.items.length).toBe(0);
-    expect(result.get("col-b" as CollectionId)?.items.length).toBe(1);
-    expect(result.get("col-b" as CollectionId)?.items[0].id).toBe("item-1");
-  });
-});
-
-describe("Pure Collision Detection Helper", () => {
-  test("detects hotspot nesting correctly when pointer is in the middle of a collection card", () => {
-    const item1 = {
-      id: "item-col-b" as TimelineItemId,
-      name: "Holiday Folder",
-      kind: "collection" as const,
-      collectionId: "col-b" as CollectionId,
-      itemCount: 0,
-      startTimeSeconds: 0,
-      durationSeconds: 10,
-    };
-
-    const itemLookup = new Map<TimelineItemId, { collectionId: CollectionId; index: number; item: TimelineItem }>([
-      ["item-col-b" as TimelineItemId, { collectionId: "col-a" as CollectionId, index: 0, item: item1 }],
-    ]);
-
-    const droppableContainers = [
-      {
-        id: "item:item-col-b",
-        rect: {
-          current: {
-            width: 200,
-            height: 100,
-            left: 100,
-            top: 100,
-            right: 300,
-            bottom: 200,
-          },
-        },
-        data: { current: {} },
-      },
-    ];
-
-    // Pointer is in the middle of card-col-b (hotspot triggers between 20% and 80%)
-    // width: 200, left: 100 -> center is 200. hotspot is 140 to 260
-    // height: 100, top: 100 -> center is 150. hotspot is 120 to 180
-    const pointerInHotspot = { x: 200, y: 150 };
-
-    const result = detectCollision({
-      active: {
-        id: "item:item-a1",
-        rect: {
-          current: {
-            width: 100,
-            height: 100,
-            left: 0,
-            top: 0,
-            right: 100,
-            bottom: 100,
-          },
-        },
-      } as any,
-      collisionRect: {
-        width: 100,
-        height: 100,
-        left: 0,
-        top: 0,
-        right: 100,
-        bottom: 100,
-      } as any,
-      droppableRects: new Map([
-        [
-          "item:item-col-b",
-          {
-            width: 200,
-            height: 100,
-            left: 100,
-            top: 100,
-            right: 300,
-            bottom: 200,
-          },
-        ],
-      ]) as any,
-      pointerCoordinates: pointerInHotspot,
-      droppableContainers: droppableContainers as any,
-      itemLookup,
-    });
-
-    expect(result.nestTargetId).toBe("col-b" as CollectionId);
-  });
-
-  test("does not trigger nesting when pointer is near the edge of a collection card", () => {
-    const item1 = {
-      id: "item-col-b" as TimelineItemId,
-      name: "Holiday Folder",
-      kind: "collection" as const,
-      collectionId: "col-b" as CollectionId,
-      itemCount: 0,
-      startTimeSeconds: 0,
-      durationSeconds: 10,
-    };
-
-    const itemLookup = new Map<TimelineItemId, { collectionId: CollectionId; index: number; item: TimelineItem }>([
-      ["item-col-b" as TimelineItemId, { collectionId: "col-a" as CollectionId, index: 0, item: item1 }],
-    ]);
-
-    const droppableContainers = [
-      {
-        id: "item:item-col-b",
-        rect: {
-          current: {
-            width: 200,
-            height: 100,
-            left: 100,
-            top: 100,
-            right: 300,
-            bottom: 200,
-          },
-        },
-        data: { current: {} },
-      },
-    ];
-
-    // Pointer is near the left edge of card-col-b (x = 110, which is outside 140 to 260)
-    const pointerNearEdge = { x: 110, y: 150 };
-
-    const result = detectCollision({
-      active: {
-        id: "item:item-a1",
-        rect: {
-          current: {
-            width: 100,
-            height: 100,
-            left: 0,
-            top: 0,
-            right: 100,
-            bottom: 100,
-          },
-        },
-      } as any,
-      collisionRect: {
-        width: 100,
-        height: 100,
-        left: 0,
-        top: 0,
-        right: 100,
-        bottom: 100,
-      } as any,
-      droppableRects: new Map([
-        [
-          "item:item-col-b",
-          {
-            width: 200,
-            height: 100,
-            left: 100,
-            top: 100,
-            right: 300,
-            bottom: 200,
-          },
-        ],
-      ]) as any,
-      pointerCoordinates: pointerNearEdge,
-      droppableContainers: droppableContainers as any,
-      itemLookup,
-    });
-
-    expect(result.nestTargetId).toBeNull();
-  });
-});
